+++ .eslintignore
@@ -1,7 +1,7 @@
 *.min.*
 frontend
 node_modules
-src/lib/markbind/src/lib/markdown-it/*
-src/lib/markbind/src/lib/markdown-it-shared/*
+packages/core/src/lib/markdown-it/*
+packages/core/src/lib/markdown-it-shared/*
 
 !.eslintrc.js
