+++ docs/devGuide/design.md
@@ -18,7 +18,7 @@ The MarkBind project is developed in a monorepo ([MarkBind/markbind](https://git
 
 * The command-line interface (CLI) application, which accepts commands from users and then uses the core library to parse and generate web pages, resides in the root.
 
-* The core library, which resolves the content include path and renders Markdown content, resides in the `src/lib/markbind/` directory.
+* The core library, which resolves the content include path and renders Markdown content, resides in the `packages/core/` directory.
 
 * The UI components library, which MarkBind authors can use to create content with complex and interactive structure, resides in the `frontend/components/` directory.
