+++ package.json
@@ -27,9 +27,9 @@
     "ciwin:all": "npm ci && npm run install:components",
     "csslint": "stylelint **/*.css",
     "csslintfix": "stylelint **/*.css --fix",
-    "install:all": "npm install && npm run install:markbind && npm run install:components",
+    "install:all": "npm install && npm run install:core && npm run install:components",
     "install:components": "cd frontend/components && npm install",
-    "install:markbind": "npm install src/lib/markbind",
+    "install:core": "npm install packages/core",
     "jest": "jest",
     "lint": "npm run lint:cli && npm run lint:components",
     "lint:cli": "eslint .",
@@ -53,6 +53,7 @@
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.7.2",
+    "@markbind/core": "file:packages/core",
     "@primer/octicons": "^9.3.1",
     "bluebird": "^3.7.2",
     "bootswatch": "^4.4.1",
@@ -71,7 +72,6 @@
     "js-beautify": "^1.10.3",
     "live-server": "^1.2.1",
     "lodash": "^4.17.15",
-    "markbind": "file:src/lib/markbind",
     "progress": "^2.0.3",
     "walk-sync": "^2.0.2",
     "winston": "^2.4.4",
