+++ packages/core/index.js
