+++ packages/core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "markbind",
+  "name": "@markbind/core",
   "version": "2.14.1",
   "description": "MarkBind core module",
   "keywords": [
@@ -15,7 +15,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/MarkBind/markbind.git",
-    "directory": "src/lib/markbind"
+    "directory": "packages/core"
   },
   "dependencies": {
     "@sindresorhus/slugify": "^0.9.1",
