+++ packages/core/src/constants.js
@@ -1,9 +1,9 @@
 module.exports = {
-  // src/lib/markbind/src/Parser.js
+  // packages/core/src/Parser.js
   ATTRIB_CWF: 'cwf',
 
   BOILERPLATE_FOLDER_NAME: '_markbind/boilerplates',
 
-  // src/lib/markbind/src/utils.js
+  // packages/core/src/utils.js
   markdownFileExts: ['md', 'mbd', 'mbdf'],
 };
