+++ packages/core/src/errors/CyclicReferenceError.js
