+++ packages/core/src/lib/markdown-it/index.js
