+++ packages/core/src/lib/markdown-it/markdown-it-block-embed/PluginEnvironment.js
