+++ packages/core/src/lib/markdown-it/markdown-it-dimmed.js
