+++ packages/core/src/lib/markdown-it/markdown-it-emoji-fixed.js
