+++ packages/core/src/lib/markdown-it/markdown-it-escape-special-tags.js
