+++ packages/core/src/lib/markdown-it/markdown-it-footnotes.js
