+++ packages/core/src/lib/markdown-it/markdown-it-icons.js
