+++ packages/core/src/lib/markdown-it/normalizeLink.js
