+++ packages/core/src/lib/markdown-it/patches/markdown-it-attrs-nunjucks.js
