+++ packages/core/src/patches/htmlparser2.js
