+++ packages/core/src/utils/nunjuckUtils.js
